+++ src/components/shared-spatial-scatterplot/AbstractSpatialOrScatterplot.js
@@ -196,7 +196,7 @@ export default class AbstractSpatialOrScatterplot extends PureComponent {
    */
   render() {
     const {
-      deckRef, viewState, uuid, layers: layerProps,
+      deckRef, viewState, uuid, layers: layerProps, hideTools,
     } = this.props;
     const { gl, tool } = this.state;
     const layers = this.getLayers();
@@ -216,9 +216,9 @@ export default class AbstractSpatialOrScatterplot extends PureComponent {
           activeTool={tool}
           setActiveTool={this.onToolChange}
           visibleTools={{
-            pan: showPanTool,
-            selectRectangle: showCellSelectionTools,
-            selectLasso: showCellSelectionTools,
+            pan: showPanTool && !hideTools,
+            selectRectangle: showCellSelectionTools && !hideTools,
+            selectLasso: showCellSelectionTools && !hideTools,
           }}
         />
         <DeckGL